Retirado redundância de função UsuariosService
